Added a new field for query arguments to the Definition named tuple
Many algorithm wrappers have two different kinds of parameter: one that
affects the construction of a data structure, and one that affects the
search strategy. By distinguishing between these two kinds of argument, the
framework will be able to reuse built data structures, which should cut
down on the amount of wasted computation. This commit adds a field for
query arguments and adds support for sending it to algorithm runner
proceses; subsequent commits will actually do something with this field