Added a weird Subprocess constructor that adds query parameter support
This "constructor" is very weird: it takes a constructor as an argument,
uses that to create a Subprocess instance, and then hastily patches that
new instance to add the set_query_arguments method. On the other hand,
query argument support doesn't fit nicely into the class hierarchy, so
maybe this is the right thing to do?